pyproject.toml
[tool.poetry]
name = "LS-AMG-RAG"
version = "0.1.0"
description = ""
authors = ["Arihant Sheth <arihant.sheth0802@gmail.com>"]
readme = "README.md"
packages = [{include = "LS_AMG_RAG"}]

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.3"
google-generativeai = "^0.4.0"
python-dotenv = "^1.0.1"
ipython = "^8.22.2"
openai = "^1.13.3"
tiktoken = "^0.6.0"
pandas = "^2.2.1"
chromadb = "^0.4.24"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
